refactor(primitives): move functions around for cohesion
