+++ crates/primitives/src/scripts/general.rs
@@ -5,7 +5,6 @@ use bitcoin::{
     transaction, Amount, OutPoint, ScriptBuf, Transaction, TxIn, TxOut, Witness,
 };
 use bitcoin_script::script;
-use bitvm::{pseudo::NMUL, treepp::*};
 use musig2::KeyAggContext;
 use secp256k1::{PublicKey, XOnlyPublicKey};
 
@@ -129,30 +128,4 @@ pub fn create_tx_outs(
             value,
         })
         .collect()
-}
-
-pub fn extract_superblock_ts_from_header() -> Script {
-    script! {
-        for i in 0..4 { { 80 - 12 + 2 * i } OP_PICK }
-        for _ in 1..4 {  { NMUL(1 << 8) } OP_ADD }
-    }
-}
-
-pub fn add_bincode_padding_bytes32() -> Script {
-    script! {
-        for b in [0; 7] { {b} } 32
-    }
-}
-
-pub fn hash_to_bn254_fq() -> Script {
-    script! {
-        for i in 1..=3 {
-            { 1 << (8 - i) }
-            OP_2DUP
-            OP_GREATERTHAN
-            OP_IF OP_SUB
-            OP_ELSE OP_DROP
-            OP_ENDIF
-        }
-    }
 }
