+++ crates/primitives/src/scripts/mod.rs
@@ -4,4 +4,3 @@ pub mod parse_witness;
 pub mod prelude;
 pub mod taproot;
 pub mod transform;
-pub mod wots;
